Refactor: Moved conversion functions to conversion_utils
